pyproject.toml
[project]
authors = [
    {name = "Aleksey_Kuznetsov", email = "aleksejkuznecov07096@gmail.com"}
]
repository = "https://github.com/Mirroel-Alvares/python-project-52/tree/main"
name = "hexlet-code"
version = "0.1.0"
description = "Task manager"
readme = "README.md"
requires-python = ">=3.11"

classifiers = [
  'Development Status :: 3 - Alpha',
  'Environment :: Web Environment',
  'Intended Audience :: End Users/Desktop',
  'Intended Audience :: Education',
  'License :: Free For Educational Use',
  'Natural Language :: English',
  'Operating System :: Microsoft',
  'Operating System :: Unix',
  'Programming Language :: Python :: 3.10',
  'Topic :: Software Development'
]

dependencies = [
    "django>=5.2",
    "django-bootstrap5>=25.1",
    "django-extensions>=4.1",
    "django-filter>=25.1",
    "python-dotenv>=1.1.0",
    "rollbar>=1.3.0",
]

[dependency-groups]
dev = [
    "coverage>=7.8.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.6",
]

[tool.hatch.build.targets.sdist]
include = ["task_manager"]

[tool.hatch.build.targets.wheel]
include = ["task_manager"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
